+++ src/main/java/dong/yoogo/domain/jira/IssueRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.time.ZonedDateTime;
 
 @Repository
-public interface IssueRepository extends JpaRepository<Issue, Long> {
+public interface IssueRepository extends JpaRepository<Issue, Integer> {
     @Query("select max(i.updated) from Issue i where i.projectKey=:pk")
     ZonedDateTime queryLastUpdatedOf(String pk);
 }
